Move all repositories to java.marvellator.repository package (only UserRepository in this case)
